Create Data.Algorand.Teal
Problem: TEAL-related data types are defined in API types module,
but they seem to be more fundamental.

Solution: create a new module in Data.Algorand and put TEAL-related
things there.
