+++ algorand-sdk.cabal
@@ -35,6 +35,7 @@ library
       Data.Algorand.Block
       Data.Algorand.MessagePack
       Data.Algorand.MessagePack.Json
+      Data.Algorand.Teal
       Data.Algorand.Transaction
       Data.Algorand.Transaction.Build
       Data.Algorand.Transaction.Group
